# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API",
        "title": "WhatAmIToDo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/api/availability/{goal_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает интервалы доступности для указанной цели",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Получить доступность по цели",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID цели",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные доступности",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAvailabilityRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid goal_id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Устанавливает интервалы доступного времени для задач указанной цели и запускает авторасписание",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Создать или обновить доступность по цели",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID цели",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные доступности",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Количество запланированных задач",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAvailabilityResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid goal_id or JSON",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/availability/{goal_id}/schedule": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Автоматически планирует \"todo\"-задачи указанной цели в доступные интервалы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Авторасписание задач по цели",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID цели",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение и число запланированных задач",
                        "schema": {
                            "$ref": "#/definitions/dto.AutoScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid goal_id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/schedule": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает запланированные задачи за день или диапазон дат",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Получить расписание",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Конкретный день YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начало диапазона YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец диапазона YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расписание за диапазон",
                        "schema": {
                            "$ref": "#/definitions/dto.GetScheduleRangeResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid date parameters",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/scheduled_tasks/{id}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переключает статус запланированной задачи (intervalID) на выполнено или нет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Отметить или отменить выполнение запланированной задачи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID запланированного задания",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Статус задачи",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ToggleTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid id or JSON",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает статистику выполненных и ожидающих задач за прошлую неделю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Получить статистику задач",
                "responses": {
                    "200": {
                        "description": "Статистика по дням",
                        "schema": {
                            "$ref": "#/definitions/dto.GetStatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tasks/upcoming": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список ближайших запланированных задач с опциональным лимитом",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Получить предстоящие задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Максимальное число задач",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUpcomingTasksResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AutoScheduleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "scheduled_tasks": {
                    "type": "integer"
                }
            }
        },
        "dto.DayAvailability": {
            "type": "object",
            "properties": {
                "day_of_week": {
                    "type": "integer"
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TimeSlotDTO"
                    }
                }
            }
        },
        "dto.DaySchedule": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ScheduledTaskDTO"
                    }
                }
            }
        },
        "dto.DayStat": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "pending": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GetScheduleRangeResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DaySchedule"
                    }
                }
            }
        },
        "dto.GetStatsResponse": {
            "type": "object",
            "properties": {
                "week": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DayStat"
                    }
                }
            }
        },
        "dto.GetUpcomingTasksResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UpcomingTaskDTO"
                    }
                }
            }
        },
...
        "dto.ScheduledTaskDTO": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "goal_title": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.TimeSlotDTO": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "dto.ToggleTaskRequest": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                }
            }
        },
        "dto.UpcomingTaskDTO": {
            "type": "object",
            "properties": {
                "goal_title": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scheduled_date": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateAvailabilityRequest": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DayAvailability"
                    }
                }
            }
        },
        "dto.UpdateAvailabilityResponse": {
            "type": "object",
            "properties": {
                "scheduled_tasks": {
                    "type": "integer"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}